feat: Add status category constants
